Add wrap_with_romanisim script/func
Completed wrap_with_romanisim python script & function. It can be run from command line or imported. Moved to py directory (so it's in the PYTHONPATH). Finally, added the wrap_with_romanisim step to mk_sim.py.
